+++ avista_sensors/sensor_sweep.py
@@ -3,9 +3,10 @@ import logging
 from threading import Thread
 import threading
 from datetime import datetime
-from avista_sensors.manager_state import ManagerState
+from avista_sensors.sweep_state import SweepState
 import avista_sensors.processor_loader as pl
 from avista_data.sensor import Sensor
+from avista_sensors.data_transporter import DataTransporter
 import RPi.GPIO as GPIO
 
 
@@ -20,12 +21,14 @@ class SensorSweep(Thread):
 
         **periodicity (int)**: periodicity of recording sensor data in milliseconds
 
+        **max_holding_period (int)**: number of sensor periods to pass through before sending data to portal
+
         **state (:obj: `ManagerState`)**: current state of the processor manager
 
         **config (dict)**: Configuration file
     """
 
-    def __init__(self, app, config=None, *args, **kwargs):
+    def __init__(self, app, periodicity, holding, config=None, *args, **kwargs):
         """Constructs a new SensorSweep instance
 
         Args:
@@ -33,17 +36,41 @@ class SensorSweep(Thread):
 
             **config (dict)**: the base configuration for the process manager
 
+            **periodicity (float)**: the number of seconds to wait between sensor recordings
+
+            **holding (int)**: number of periods of data to store prior to sending to the server
+
             __*args__: variable args passed to super class
 
             __**kwargs__: keyword args passed to super class
         """
         super(SensorSweep, self).__init__(*args, **kwargs)
         self.processors = []
-        self.periodicity = 1
-        self.state = ManagerState.IDLE
+        self.periodicity = periodicity
+        self.max_holding_period = holding
+        self.state = SweepState.IDLE
         self.config = config
         self._kill = threading.Event()
+        self._transporter = DataTransporter()
         self.app = app
+
+    def set_periodicity(self, periodicity):
+        if float(periodicity) and periodicity > 0:
+            self.periodicity = periodicity
+        else:
+            raise Exception("periodicity must be a float greater than 0")
+
+    def get_periodicity(self):
+        return self.periodicity
+
+    def set_holding_period(self, holding):
+        if int(holding) and holding > 0:
+            self.max_holding_period = holding
+        else:
+            raise Exception("holding period must be an integer greater than 0")
+
+    def get_holding_period(self, holding):
+        return self.max_holding_period
 
     def run(self):
         """Main execution body of the thread"""
@@ -52,19 +79,19 @@ class SensorSweep(Thread):
 
     def _begin(self):
         """Starts the sensor sweep"""
-        self.state = ManagerState.STARTING
+        self.state = SweepState.STARTING
 
         logging.info("Sensor Sweep Starting")
 
     def stop(self):
         """Stops the sensor sweep by setting the _kill event"""
-        self.state = ManagerState.STOPPING
+        self.state = SweepState.STOPPING
 
         logging.info("Sensor Sweep Stopping")
 
         GPIO.cleanup()
 
-        self.state = ManagerState.IDLE
+        self.state = SweepState.IDLE
         self._kill.set()
 
     def stopped(self):
@@ -77,7 +104,7 @@ class SensorSweep(Thread):
 
     def init(self):
         """Initializes the processors using a processor loader"""
-        self.state = ManagerState.INITIALIZING
+        self.state = SweepState.INITIALIZING
         logging.info("Sensor Sweep Initializing")
 
         GPIO.setmode(GPIO.BCM)
@@ -93,10 +120,11 @@ class SensorSweep(Thread):
         Here each attached sensor is polled, the process sleeps for the specified periodicity then continues.
         This process goes on until the sensor sweep is stopped.
         """
-        self.state = ManagerState.EXECUTING
+        self.state = SweepState.EXECUTING
 
         logging.info("Sensor Sweep Running")
 
+        periods = 0
         while True:
             with self.app.app_context():
                 if self.stopped():
@@ -105,3 +133,6 @@ class SensorSweep(Thread):
                     if p is not None:
                         p.process(int(datetime.timestamp(datetime.now())))
                 time.sleep(self.periodicity)
+                periods += 1
+                if periods >= self.max_holding_period:
+                    self._transporter.transfer()